ai-service/requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
spacy
pydantic
python-multipart
